finalized code, created CommonVerifyers for common verify actions created WholeListSteps steps file that contains steps related doing tests on the whole list of articles
