+++ src/test/java/steps/ArticleDetailSteps.java
@@ -25,7 +25,6 @@ public class ArticleDetailSteps {
         selectedDetailPage = new ArticleDetailPage(driver);
     }
 
-
     @When("user clicks on the share button")
     @Step("user clicks on the share button")
     public void userClicksOnTheShareButton() {
@@ -35,36 +34,29 @@ public class ArticleDetailSteps {
     @Then("the app should switch to gmail")
     @Step("the app should switch to gmail")
     public void theAppShouldSwitchToGmail() {
-        String welcomeTourActivity = ".welcome.WelcomeTourActivity";
-        String composeActivity = ".ComposeActivityGmailExternal";
+        String composeActivity = ".ComposeActivityGmail";
+        String composeActivityExternal = ".ComposeActivityGmailExternal";
 
         String currentActivity = gmailPage.getCurrentActivity();
 
-        Assert.assertTrue( currentActivity.equals(welcomeTourActivity)|| currentActivity.equals(composeActivity));
+        Assert.assertTrue( currentActivity.equals(composeActivityExternal)|| currentActivity.equals(composeActivity), "Current app activity is not a compose activity" + driver.currentActivity());
     }
 
     @When("user checks the name of the author and the title")
     @Step("user checks the name of the author and the title")
     public void userChecksTheNameOfTheAuthorAndTheTitle() {
-        this.authorEmail = selectedDetailPage.getAuthorEmail();
         this.articleTitle = selectedDetailPage.getTitle();
+        this.authorEmail = selectedDetailPage.getAuthorEmail();
     }
 
     @And("the recipient and subject fields are filled")
     @Step("the recipient and subject fields are filled")
     public void theRecipientAndSubjectFieldsAreFilled() {
-
-        Assert.assertEquals(gmailPage.getCurrentActivity(), ".ComposeActivityGmailExternal", "User is not in compose email activity.");
-
-        //removing "author: " here as it can be needed on other places
-        String expectedRecipientText = "<" + authorEmail.replace("author: ", "") + ">, ";
-        Assert.assertEquals(gmailPage.getRecipient(), expectedRecipientText, "Testing if actual recipient equals expected recipient");
-        Assert.assertEquals(gmailPage.getSubject(), articleTitle, "Testing if subject equals to the title of the article");
-        Assert.assertTrue(gmailPage.getBody().isEmpty(), "Testing if email body is empty");
+        CommonVerifyers.verifyGmailFilledCorrectly(authorEmail,articleTitle, gmailPage);
     }
 
-
     @And("user navigates back to viesure application")
+    @Step("user navigates back to viesure application")
     public void userNavigatesBackToViesureApplication() {
         //First one closes the keyboard, second navigates back
         driver.navigate().back();
